Retry on 429 response code (throttling)
The end2end tests evidenced that this can happen, despite the shimmy
being a single process. In this case, we stop processing the message and
let it go back to the queue for a retry.

Testing the logic with unit tests, more difficult to actually test the
integration with SQS. There may be a better way of triggering a retry,
and this logic targets 10 seconds from the 429 error.