First part of bot communication with lax:
PublishPerfectArticle is split in two, IngestArticleZip does the same things up to ExpandArticle, then a new activity, IngestToLax writes a message in a queue that will be consumed by lax.
Coming Next: adapter in bot will consume lax response that will be written to another queue. This will trigger the second part of the old PublishPerfectArticle workflow (from ApplyVersionNumber).