Logging handler compatible with logrotate
logrotate is a Unix daemon that every morning takes the log file, compresses it as worker.log.1.gz, and then creates a new empty one to be written by the application.

The Python code needs to be aware of this to reopen the file after it has been changed, otherwise logs written after the rotation will be lost.