copies contents added to elife-publishing-cdn to elife-published/articles
plus
Fixes content-type download problem: When loading a key from S3, only a flat copy is made and we don’t have access to the current metadata in this copy, so content-type was being lost. Now I add the content-type when adding content-disposition.

The copy is made in the following pattern

elife-published/articles/[article id]