Prevent multiple workflows running concurrently
AFAIK, two workflows with the same id will be stopped from running (the
second one will be dropped). For these workflows it seems safe to say we
don't want more than one copy running at a given time.

The pid remains useful for the identity of processes (like worker.py),
which will continue to log it. It's also logged as the `Identity` field
of all the activities on the SWF console, as different activities are
performed by different workers (e.g. `worker_31510`)