VerifyImageServer should only check the .tif version of a figure
We are falling back to the .jpg version if the .tif doesn't work into the IIIF server, but this is done transparently by the server itself. So I believe we don't need to also check the .jpg, and we can do half the work that this activity is doing for better reliability and speed.