Allowing linting on activities that use StorageContext
S3StorageContext is rightly behind an indirection, but client code instantiates StorageContext and gets an instance of another class instead, S3StorageContext itself. This makes the linting go crazy as it sees method of S3StorageContext being called on what it believes is just a StorageContext. The classes have no relation with each other, so we provide a factory function here that does the same job but in a simpler way.